+++ inc/unitmodules.hpp
@@ -173,8 +173,6 @@ public:
 
     // Is simulation data stored to memory.
     void Set_EnableStore(bool store=false);
-    // Is simulation data printed to a file.
-    void Set_EnablePrint(bool print=true);
 
     // all input data will be multiplied by "inputgain" before stored.
     void Set_InputGain(double inputgain=1);
@@ -190,9 +188,9 @@ private:
     // Stored data.
     std::vector<double> _values;
 
-    // See public member function "Set_EnableStore" and "Set_EnablePrint".
-    // @_maxstorage: How many samples will it store.
-    bool _store, _print;
+    // See public member function "Set_EnableStore".
+    bool _store;
+    // _maxstorage: How many samples will it store.
     int _maxstorage;
 
     // @_outvalue: value of its latest sample point.